gdf2 working (without subsambling and WITHOUT EVENTS)

+++ src/base/signal_channel.cpp
@@ -11,7 +11,7 @@ SignalChannel::SignalChannel(uint32 number, const QString& label,
                              const QString& physical_dim,
                              float64 physical_minimum,
                              float64 physical_maximum,
-                             int64 digital_minimum, int64 digital_maximum,
+                             float64 digital_minimum, float64 digital_maximum,
                              uint32 data_type, uint32 data_offset,
                              const QString filter_label, float64 lowpass,
                              float64 highpass, bool notch)
